Move common files into their own directory
It is likely that there will be many common files between
hardware-specific configurations. We move the generic files into a
common directory, which is synced into the root of the Zip file before
having machine-specific files synced over the top. We can override any
of these common files by creating a file of the same name in the
machine-specific directory.

Signed-off-by: Ben Agricola <ben+git@agrico.la>