Update LDO kit toolsetter pin
Our initial configuration during Dave's testing used the `PA9` pin for
the toolsetter due to it not working with the `zmax` port unless the
correct jumper is in place to set all the endstops to 5v.

Adding the jumper for 5v is in the LDO kit instructions so we should
use the intended pin for our config.

Signed-off-by: Ben Agricola <ben+git@agrico.la>