Switch to pin ids
There was a mismatch between the teamgloomy documentation and firmware
pin aliases, and also some shenanigans with our internal testing
regarding 5v endstop ports that meant the Spindle RPM pin was set to
`PE6`. This is important, because the LDO documentation wires the
spindle RPM to the "Neopixel" port (`PD_15`).

This is problematic for 2 reasons - the first is that this port
automatically changes state based on RRF's state, and is directly linked
to the onboard status LED as well. The second is that this port is
pulled high in hardware, so under a failure situation it _could_ be
pulled high and start the spindle.

We don't want this to happen, and we also don't want to disable the
status LED.

So the Spindle pin definition has not changed (except to consistently
match the format `P[A-I]_[0-9]+`).

We also add a warning to `M552.1` about not using it from `config.g`.

Signed-off-by: Ben Agricola <ben+git@agrico.la>