Build multiple machine configs
To support multiple machine types and configs, this modifies the release
script so it takes a machine type and id as first and second variables.

Environment variables are loaded for the machine type and the machine ID
is used to pick the source folder for the config files.

We also modify the GitHub workflow to call the release script correctly.

Signed-off-by: Ben Agricola <ben+git@agrico.la>