Implement optional example files
Previously, it has not been clear that the toolsetter.g and touchprobe.g
files were boilerplate and the user should edit those files to configure
the feature if it does not come as standard on their machine build.

To avoid this, we now load the estop, toolsetter and touchprobe files
conditionally (if they exist) - and only package these with a
machine-specific configuration if they are actually implemented on the
machine. We package example files, with the suffix ".example" as part of
the common configuration, and if the matching 'concrete' file (without
the ".example" suffix) exists in the machine-specific configuration then
we delete the example file.

Signed-off-by: Ben Agricola <ben+git@agrico.la>